NORM appealed

+++ inc/cub3d_keys.h
@@ -15,38 +15,38 @@
 
 # ifdef __APPLE__
 
-#  define ESC_KEY		(int)53
+#  define ESC_KEY		53
 
-#  define W_KEY			(int)13
-#  define A_KEY			(int)0
-#  define S_KEY			(int)1
-#  define D_KEY			(int)2
+#  define W_KEY			13
+#  define A_KEY			0
+#  define S_KEY			1
+#  define D_KEY			2
 
-#  define UP_KEY		(int)126
-#  define DWN_KEY		(int)125
-#  define RGT_KEY		(int)124
-#  define LFT_KEY		(int)123
+#  define UP_KEY		126
+#  define DWN_KEY		125
+#  define RGT_KEY		124
+#  define LFT_KEY		123
 
-#  define ZOOM_IN_KEY	(int)4
-#  define ZOOM_OUT_KEY	(int)5
+#  define ZOOM_IN_KEY	4
+#  define ZOOM_OUT_KEY	5
 
-#  define SPC_KEY		(int)49
-#  define ENT_KEY		(int)36
-#  define QUE_KEY		(int)12
+#  define SPC_KEY		49
+#  define ENT_KEY		36
+#  define QUE_KEY		12
 
 # elif __linux__
 
-#  define ESC_KEY			(int)65307
+#  define ESC_KEY			65307
 
-#  define A_KEY				(int)97
-#  define D_KEY				(int)100
-#  define W_KEY       		(int)119
-#  define S_KEY				(int)115
+#  define A_KEY				97
+#  define D_KEY				100
+#  define W_KEY       		119
+#  define S_KEY				115
 
-#  define LFT_KEY			(int)65361
-#  define RGT_KEY			(int)65363
-#  define UP_KEY			(int)65362
-#  define DWN_KEY			(int)65364
+#  define LFT_KEY			65361
+#  define RGT_KEY			65363
+#  define UP_KEY			65362
+#  define DWN_KEY			65364
 
 // #  define ZOOM_IN_KEY	(int)4
 // #  define ZOOM_OUT_KEY	(int)5
